pyproject.toml
[project]
name = "canvas-mcp"
version = "0.1.0"
description = "Canvas MCP server with FastMCP tools"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "requests",
    "fastmcp",  # make sure this is the correct package name for FastMCP
    "beautifulsoup4",
]

[project.scripts]
canvas-mcp = "canvas_mcp.run:main"


[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]